Refactor builder to keep only one function per file
