--- builder/src/get-temp-dir.ts
@@ -1,9 +1,0 @@
-import { mkdtempSync } from 'fs';
-import { tmpdir } from 'os';
-import { join } from 'path';
-
-const tempDirPath = mkdtempSync(join(tmpdir(), 'walden-'));
-
-export default function getTempDir(): string {
-  return tempDirPath;
-}
